sprint 12: 17) battle, 18) Stats

+++ src/main/java/org/mtcg/app/models/Card.java
@@ -13,7 +13,7 @@ public class Card
     @JsonProperty("Name")
     private String name;
     @JsonProperty("Damage")
-    private double damage;
+    private int damage;
 
 
     public Card()
@@ -22,7 +22,7 @@ public class Card
 
 
 
-    public Card(String id, String name, double damage)
+    public Card(String id, String name, int damage)
     {
         this.id = id;
         this.name = name;
