+++ src/main/java/org/mtcg/app/models/MonsterCard.java
@@ -2,7 +2,7 @@ package org.mtcg.app.models;
 
 public class MonsterCard extends Card
 {
-    public MonsterCard(String id, String name, double damage)
+    public MonsterCard(String id, String name, int damage)
     {
         super(id, name, damage);
     }
