+++ src/main/java/org/mtcg/app/models/SpellCard.java
@@ -2,7 +2,7 @@ package org.mtcg.app.models;
 
 public class SpellCard extends Card
 {
-    public SpellCard(String id, String name, double damage)
+    public SpellCard(String id, String name, int damage)
     {
         super(id, name, damage);
     }
